+++ src/rest.ts
@@ -2,5 +2,5 @@ export {ResponseType, LocationHeaderAndJsonResponse, LocationHeaderResponse, Blo
 export {Cache} from './rest/cache';
 export {Body,BaseUrl,DefaultHeaders,DELETE,GET,HEAD,Header,Headers,JsonContentType,ParameterTransform,Path,POST,Produces,PUT,Query,QueryObject,ResponseTransform,RESTClient,FullHttpResponse,ReportProgress, DisableInterceptor, ProgressIndicatorGroup} from './rest/common';
 export {RestTransferStateService} from './transferState/restTransferState.service';
-export {AdditionalInfoPropertyDescriptor} from './rest/rest.interface';
+export {AdditionalInfoPropertyDescriptor, ParametersMetadata, RestCaching, RestDisabledInterceptors, RestFullHttpResponse, RestHttpHeaders, RestMethod, RestParameters, RestReportProgress, RestResponseTransform, RestResponseType, KeyIndex, ParametersTransformMetadata} from './rest/rest.interface';
 export {HTTP_HEADER_ACCEPT, HTTP_HEADER_CONTENT_TYPE} from './rest/constants';